# api/apps/api/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MarxanCloud API",
    "description": "MarxanCloud is a conservation planning platform.",
    "version": "0.4.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
        }
    },
    "schemas": {
...
      "ProtectedArea": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "wdpaId": {
            "type": "number"
          },
          "fullName": {
            "type": "string"
          },
          "iucnCategory": {
            "type": "string"
          },
          "shapeLength": {
            "type": "number"
          },
          "shapeArea": {
            "type": "number"
          },
          "countryId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          },
          "theGeom": {
            "type": "object"
          }
        },
        "required": ["id", "wdpaId"]
      },
      "JSONAPIProtectedAreaData": {
        "type": "object",
        "properties": {
          "type": {
            "type": "object"
          },
          "id": {
            "type": "string"
          },
          "attributes": {
            "$ref": "#/components/schemas/ProtectedArea"
          }
        },
        "required": ["type", "id", "attributes"]
      },
      "ProtectedAreaResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JSONAPIProtectedAreaData"
          }
        },
        "required": ["data"]
      },
      "IUCNProtectedAreaCategoryDTO": {
        "type": "object",
        "properties": {
          "iucnCategory": {
            "type": "string"
          }
        }
      },
      "JSONAPIIUCNProtectedAreaCategoryData": {
        "type": "object",
        "properties": {
          "type": {
            "type": "object"
          },
          "id": {
            "type": "string"
          },
          "attributes": {
            "$ref": "#/components/schemas/IUCNProtectedAreaCategoryDTO"
          }
        },
        "required": ["type", "id", "attributes"]
      },
      "IUCNProtectedAreaCategoryResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JSONAPIIUCNProtectedAreaCategoryData"
          }
        },
        "required": ["data"]
      }
    }
  },
  "paths": {
...
    "/api/v1/protected-areas": {
      "get": {
        "operationId": "ProtectedAreasController_findAll",
        "summary": "",
        "description": "Find all protected areas",
        "parameters": [
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "A comma-separated list of relationship paths. Allows the client to customize which related resources should be returned.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "An array of filters (e.g. `filter[keyA]=<value>&filter[keyB]=<value1>,<value2>...`). Allows the client to request for specific filtering criteria to be applied to the request. Semantics of each set of filter key/values and of the set of filters as a whole depend on the specific request. Available filters: `fullName`, `wdpaId`, `iucnCategory`, `status`, `designation`, `countryId`.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "A comma-separated list that refers to the name(s) of the fields to be returned. An empty value indicates that all fields will be returned (less any fields specified as `omitFields`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "omitFields",
            "required": false,
            "in": "query",
            "description": "A comma-separated list that refers to the name(s) of fields to be omitted from the results. This could be useful as a shortcut when a specific field such as large geometry fields should be omitted, but it is not practical or not desirable to explicitly whitelist fields individually. An empty value indicates that no fields will be omitted (although they may still not be present in the result if an explicit choice of fields was provided via `fields`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "A comma-separated list of fields of the primary data according to which the results should be sorted. Sort order is ascending unless the field name is prefixed with a minus (for descending order).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page[size]",
            "required": false,
            "in": "query",
            "description": "Page size for pagination. If not supplied, pagination with default page size of 25 elements will be applied.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page[number]",
            "required": false,
            "in": "query",
            "description": "Page number for pagination. If not supplied, the first page of results will be returned.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "disablePagination",
            "required": false,
            "in": "query",
            "description": "If set to `true`, pagination will be disabled. This overrides any other pagination query parameters, if supplied.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtectedAreaResult"
                }
              }
            }
          }
        },
        "tags": ["ProtectedArea"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/protected-areas/preview/tiles/{z}/{x}/{y}.mvt": {
      "get": {
        "operationId": "ProtectedAreasController_proxyProtectedAreaTile",
        "summary": "",
        "description": "Get tile for protected areas.",
        "parameters": [
          {
            "name": "bbox",
            "required": false,
            "in": "query",
            "description": "Bounding box of the project [xMin, xMax, yMin, yMax]",
            "example": [-1, 40, 1, 42],
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "Id of WDPA area",
            "example": "e5c3b978-908c-49d3-b1e3-89727e9f999c",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "required": true,
            "in": "path",
            "description": "The tile y offset on Mercator Projection",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x",
            "required": true,
            "in": "path",
            "description": "The tile x offset on Mercator Projection",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "z",
            "required": true,
            "in": "path",
            "description": "The zoom level ranging from 0 - 20",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/"
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "tags": ["ProtectedArea"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/protected-areas/iucn-categories": {
      "get": {
        "operationId": "ProtectedAreasController_listIUCNProtectedAreaCategories",
        "summary": "",
        "description": "Find unique IUCN categories among protected areas in a single given administrative area.",
        "parameters": [
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "A comma-separated list of relationship paths. Allows the client to customize which related resources should be returned.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "An array of filters (e.g. `filter[keyA]=<value>&filter[keyB]=<value1>,<value2>...`). Allows the client to request for specific filtering criteria to be applied to the request. Semantics of each set of filter key/values and of the set of filters as a whole depend on the specific request.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "A comma-separated list that refers to the name(s) of the fields to be returned. An empty value indicates that all fields will be returned (less any fields specified as `omitFields`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "omitFields",
            "required": false,
            "in": "query",
            "description": "A comma-separated list that refers to the name(s) of fields to be omitted from the results. This could be useful as a shortcut when a specific field such as large geometry fields should be omitted, but it is not practical or not desirable to explicitly whitelist fields individually. An empty value indicates that no fields will be omitted (although they may still not be present in the result if an explicit choice of fields was provided via `fields`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "A comma-separated list of fields of the primary data according to which the results should be sorted. Sort order is ascending unless the field name is prefixed with a minus (for descending order).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page[size]",
            "required": false,
            "in": "query",
            "description": "Page size for pagination. If not supplied, pagination with default page size of 25 elements will be applied.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page[number]",
            "required": false,
            "in": "query",
            "description": "Page number for pagination. If not supplied, the first page of results will be returned.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "disablePagination",
            "required": false,
            "in": "query",
            "description": "If set to `true`, pagination will be disabled. This overrides any other pagination query parameters, if supplied.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "filter[adminAreaId]",
            "required": true,
            "in": "query",
            "description": "Only protected areas within the given admin area will be considered.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IUCNProtectedAreaCategoryResult"
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "tags": ["ProtectedArea"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/protected-areas/{id}": {
      "get": {
        "operationId": "ProtectedAreasController_findOne",
        "summary": "",
        "description": "Get protected area by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "A comma-separated list of relationship paths. Allows the client to customize which related resources should be returned.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "A comma-separated list that refers to the name(s) of the fields to be returned. An empty value indicates that all fields will be returned (less any fields specified as `omitFields`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "omitFields",
            "required": false,
            "in": "query",
            "description": "A comma-separated list that refers to the name(s) of fields to be omitted from the results. This could be useful as a shortcut when a specific field such as large geometry fields should be omitted, but it is not practical or not desirable to explicitly whitelist fields individually. An empty value indicates that no fields will be omitted (although they may still not be present in the result if an explicit choice of fields was provided via `fields`).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtectedAreaResult"
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "tags": ["ProtectedArea"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  }
}